Fixes the performance issue in `Project()` (#2383)
## What do these changes do?

`ProjectToSimple()` requires scan all edges to establish the new offset
ranges. The previous implements runs slow (slower than scan edges for
many times in pagerank) because

- It use a scan implementation to find the start and end point for a
given label. As the CSR is internally sorted, a bisect could be enough,
the complexity is `O(E)`.
- It use a single-thread implementation for vertices
- The method `getRangeOfLabel` takes `std::shared_ptr` as argument, the
shared ptr would be constructed and destructed for `O(V)` times.

This PR optimize the implementation. Benchmark shows that it improve the
`Project()` operation from `2~3 seconds` to `0.02~0.1 seconds` (consider
the overhead of creating threads for parallelism) on an internal
datasets. (PageRank requires about `0.02~0.03 seconds` to scan the
edges).

## Related issue number

Part of #1300

Signed-off-by: Tao He <linzhu.ht@alibaba-inc.com>