Fix: Gremlin failed to start when k8s release name contains' - '. #2524 (#2526)
## What do these changes do?

<!-- Please give a short brief about these changes. -->
When starting the gremlin instance, you need to get the instance_id. If
the release name of the current application contains' - ', for example,
'b-c', the coordinator_name is' coordinator-b-c ', and every char after
the last '-' is captured in the case of existing code, and the
instance_id is' c '. In fact, every char after the first '-' should be
captured. instance_id is' b-c '.

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this
change? -->
Fixes #2524