[GIE] Support `Until` condition in `PathExpand` (#2514)
<!--
Thanks for your contribution! please review
https://github.com/alibaba/GraphScope/blob/main/CONTRIBUTING.md before
opening an issue.
-->

## What do these changes do?

<!-- Please give a short brief about these changes. -->

Support `UNTIL` condition in `PathExpand`. e.g., you can query like:

1） `g.V().hasLabel('person').has('name',
'marko').out("0..10").with('UNTIL', '@.name=='josh'')`,
which means to find the path starting from 'marko', and ends if the
target vertex is 'josh', or the path's length exceeds 10.

2） `g.V().hasLabel('person').has('name',
'marko').out("5..10").with('UNTIL', '@.name=='josh'')`,
which means to find the path starting from 'marko', the path's length
must be no less than 5, and ends either the target vertex is 'josh', or
the path's length exceeds 10.

3） `g.V().hasLabel('person').has('name',
'marko').out("1..10").with('UNTIL', '@.name=='marko'')` ,
which means to find the path starting from 'marko', and ends if the
target vertex is 'marko' (i.e., a cycle detection), or the path's length
exceeds 10. NOTICE that in this case, DO NOT give the lower bound as
`0`. If `0` is given, the path will stop immediately at the starting
'marko' vertex since it already meets the until condition.


## Related issue number

<!-- Are there any issues opened that will be resolved by merging this
change? -->

Fixes #2516

---------

Co-authored-by: shirly121 <yihe.zxl@alibaba-inc.com>
Co-authored-by: Longbin Lai <longbin.lailb@alibaba-inc.com>