Allow finetune parameters for each statefulsets (#2625)
This pull request improves the customizability and usability of the
GraphScope Store Helm chart by using more specific values for different
pod settings in the `statefulset.yaml` files of each component. It also
fixes some minor typos in the `values.yaml` file, the `NOTES.txt` file,
and the `IngestProcessor` and `BatchSender` classes.