Use vineyardctl API to inject the vineyard sidecar (#2612)
This pull request refactors the vineyard deployment in GraphScope to use
a sidecar container instead of a separate container in the engine pod.
This improves the performance and stability of the system and simplifies
the management of vineyard. It also fixes some typos and errors in the
documentation. The main changes are in
`coordinator/gscoordinator/cluster_builder.py` and
`coordinator/gscoordinator/kubernetes_launcher.py`.

---------
Signed-off-by: Ye Cao <caoye.cao@alibaba-inc.com>