[GIE Compiler] Fix query params of getV base in path expand operator (#2773)
<!--
Thanks for your contribution! please review
https://github.com/alibaba/GraphScope/blob/main/CONTRIBUTING.md before
opening an issue.
-->

## What do these changes do?
In the cypher statement `(a:person)-[b:knows]-(c:person)` that
represents a path expansion, the logical plan includes two `getV`
operators. One is `getV` base inner the `path_expand` b, another is the
`endV` operator following the b. It's important to note that these
operators have different query parameters.

The `getV` base can retrieve vertices of any type that are adjacent to
the type specified in the `path_expand` operation. **This pr
specifically addresses this requirement**. The `endV` operator is
constrained by the type specified in the given query.
<!-- Please give a short brief about these changes. -->

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this
change? -->

Fixes

Co-authored-by: Longbin Lai <longbin.lailb@alibaba-inc.com>