[GIE] Fix unbalanced data routing in DefaultRouter in GIE (#2994)
As titled.

After refine the `DefaultRouter` implementation, we can see that all
workers are doing the computation, which also brings about 20%
acceleration in total time cost.

```
job:  1
worker: 0
computation cost is 417248.23 ms, and the ratio of computation cost to total cost is 0.85
worker: 1
computation cost is 434071.14 ms, and the ratio of computation cost to total cost is 0.88
worker: 2
computation cost is 430534.29 ms, and the ratio of computation cost to total cost is 0.88
worker: 3
computation cost is 434731.57 ms, and the ratio of computation cost to total cost is 0.88
...
```

Committed-by: bingqing.lbq from Dev container

<!--
Thanks for your contribution! please review
https://github.com/alibaba/GraphScope/blob/main/CONTRIBUTING.md before
opening an issue.
-->

## What do these changes do?

<!-- Please give a short brief about these changes. -->

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this
change? -->

Fixes #2993