fix(interactive): Fix the bug in the RexNode to Proto Conversion (#3171)
<!--
Thanks for your contribution! please review
https://github.com/alibaba/GraphScope/blob/main/CONTRIBUTING.md before
opening an issue.
-->

## What do these changes do?
For a specific expression, i.e. `a.age-(b.age-1)`, there is an issue
when converting it to a proto structure, resulting in missing
parentheses in the generated proto; it gets converted to
`a.age-b.age-1`. This PR is meant to address the issue.

<!-- Please give a short brief about these changes. -->

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this
change? -->

Fixes