fix(interactive): Fix bugs in Label Name to Id Conversion (#3221)
<!--
Thanks for your contribution! please review
https://github.com/alibaba/GraphScope/blob/main/CONTRIBUTING.md before
opening an issue.
-->

## What do these changes do?
When converting a label name into an ID, it is necessary to first infer
whether it is a vertex label or an edge label based on the operator
preceding the elementMap. There will be issues when the preceding
operator falls into one of the following categories:
1. When the preceding operator is an `ExpandFusionStep` after
optimization (fusion of outE + filter + inV);
2. When the preceding operator is an operator like `limit` or `where`
that does not change the input type.

This PR addresses the two issues mentioned above.

<!-- Please give a short brief about these changes. -->

## Related issue number
fix #3219 

<!-- Are there any issues opened that will be resolved by merging this
change? -->

Fixes

Co-authored-by: Longbin Lai <longbin.lailb@alibaba-inc.com>