fix(interactive): Distinguish 'label equals' or 'unique key equals' from Filter Conditions (#3285)
<!--
Thanks for your contribution! please review
https://github.com/alibaba/GraphScope/blob/main/CONTRIBUTING.md before
opening an issue.
-->

## What do these changes do?
1. distinguish 'label equals' or 'unique key equals' from filter
conditions, 'label equals' is fused to decide the label type while
'unique key equals' is fused for index predicate.
2. convert index predicate from RelNode to ir core.
3. handle special cases when converting 'SEARCH' operator from RelNode
to ir core. Specifically, if 'SEARCH' happens in index predicate or the
operand of 'SEARCH' is a continuous range, it will be converted to
multiple conditions connected by 'and' or 'or'. These conversions may
not necessarily be reasonable, at least in order to make ir core
supported.
<!-- Please give a short brief about these changes. -->

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this
change? -->

Fixes

---------

Co-authored-by: Longbin Lai <longbin.lailb@alibaba-inc.com>