fix(interactive): Fix bugs of Groot Bulk Load When Retrying to Download Source Files (#3438)
Due to the instability of the network and tunneling tools, a certain
store encountered an error while downloading the original file. At this
point, the file already exists but only a portion has been downloaded,
causing subsequent retry operations to result in a `FileAlreadyExisted`
error. This PR addresses the issue by enabling the `REPLACE_EXISTING`
option, allowing the overwriting of existing files.