feat(flex): Configure the memory strategy with memory-level. (#3511)
Passing parameter "--memory-level=n" to rt_server to configure memory
allocation strategy.
- n = 0: memory allocation will be mmap to files in SHARED mode.
Suitable for scenarios with severe memory constraints.
- n = 1: memory allocation will be mmap to files in PRIVATE mode. The
modified parts in the storage cannot be swapped to disk; if the modified
portions become too large, it may lead to a crash due to insufficient
memory.
- n = 2: the storage of certain data will utilize hugepages, which can
enhance the system's throughput. However, the portion using hugepages
will be fixed in memory usage, and hugepages require static allocation.
- n = 3: the storage of all data will endeavor to make use of hugepages
as much as possible.