fix(interactive): Fix Bug in Operator Precedence of `not` (#3598)
<!--
Thanks for your contribution! please review
https://github.com/alibaba/GraphScope/blob/main/CONTRIBUTING.md before
opening an issue.
-->

## What do these changes do?
The operator precedence of `not` in sql is low, which is the same as
`and` or `or`. However, the `not` operator is implemented by `!` in c++
or in rust, which precedence is high. To solve the inconsistency, we
always add brace for operand of `not` operator.

<!-- Please give a short brief about these changes. -->

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this
change? -->

Fixes #3601

---------

Co-authored-by: Zhang Lei <xiaolei.zl@alibaba-inc.com>