fix(interactive): Migrating Physical Layer (Repartition) into New Compilation Stack (#3586)
<!--
Thanks for your contribution! please review
https://github.com/alibaba/GraphScope/blob/main/CONTRIBUTING.md before
opening an issue.
-->

## What do these changes do?

<!-- Please give a short brief about these changes. -->

As titled. This pr deals with repartition in physical layer as listed in
#3561:
1. Add repartition operator to visit remote vertices/edges.
2. Support two strategies for property fetching: 1) for vertex property,
we fetch it lazily when using it; 2) for edge property, we precache it
once we visit the edges. The main reason is that the store does not
support to get edge properties by edge id, so we have to cache it when
visiting edges.

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this
change? -->

#3561

---------

Co-authored-by: shirly121 <yihe.zxl@alibaba-inc.com>