fix(interactive): Fix CI Bug Caused By PR #3630 (#3647)
<!--
Thanks for your contribution! please review
https://github.com/alibaba/GraphScope/blob/main/CONTRIBUTING.md before
opening an issue.
-->

## What do these changes do?
The issue is caused by PR #3630, which changed the result collection
from asynchronous to synchronous. After submitting a request to the
client, it is necessary to actively call a request to retrieve the
result from the buffer. However, the test chain forgot to add this call,
leading to the problem. This PR mainly fix the issue.

<!-- Please give a short brief about these changes. -->

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this
change? -->

Fixes