fix(interactive): Introduce the high-order statistics provider `GLogue` into the new Compilation Stack (#3609)
<!--
Thanks for your contribution! please review
https://github.com/alibaba/GraphScope/blob/main/CONTRIBUTING.md before
opening an issue.
-->

## What do these changes do?

<!-- Please give a short brief about these changes. -->

This PR introduces `GLogue`, which precomputed and preserves the frequency of small graph patterns (a.k.a., motifs) with size up to $k$ in the data graph as high-order statistics.
In addition, the PR encompasses the foundational structure of `Pattern` and several functions reliant on `Pattern`, such as isomorphism checking, among others.

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this
change? -->

Fixes #3657 

---------

Co-authored-by: shirly121 <yihe.zxl@alibaba-inc.com>