fix(interactive): Fix Bug of Duplicated Plan Id Generation After Failover (#3710)
<!--
Thanks for your contribution! please review
https://github.com/alibaba/GraphScope/blob/main/CONTRIBUTING.md before
opening an issue.
-->

## What do these changes do?
Re-implement `QueryIdGenerator` strategy to avoid duplicated plan id
generation after failover. The new implemented function is to assign a
unique job id for the query, which needs to ensure uniqueness in time
and space.

Specifically, for queries a and b, the following two cases are required
not to produce duplicated plan ids:
1. When query a and b are sent to the same frontend but at different
times;
2. When query a and b are sent to different frontends, regardless if the
timestamps are the same or not;

This requires us to implement the functionality of `UUID`. The default
Java `UUID` is composed of 128 bits, with the first 64 bits determined
by timestamp and the latter 64 bits composed of MAC address plus other
items. However, our job id is made up of 64 bits, and converting 128
bits into 64 bits might compromise the uniqueness of the id;

Therefore, we adopt another method similar to random number generation.
We use Java’s built-in `SecureRandom`, which can directly generate a
random 64-bit id through `SecureRandom#nextLong()`. `SecureRandom` is a
method of pseudo-random generation, that is, deterministic generation of
a random sequence with a random seed determined by a combination of
timestamp + machine physical properties, which basically meets our needs
for plan id.

<!-- Please give a short brief about these changes. -->

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this
change? -->

Fixes #3702