fix(interactive): Fix Bugs of `Where Subquery` in Gremlin (#3757)
<!--
Thanks for your contribution! please review
https://github.com/alibaba/GraphScope/blob/main/CONTRIBUTING.md before
opening an issue.
-->

## What do these changes do?
fix bugs of `where subquery` in gremlin:
```
g.V().as("a").out().as("b").select("a", "b").where(__.as("a").out("knows").as("b"))
g.V().as("a").out().in().as("b").select("a", "b").where(__.as("b").has("name", "marko"))
```

<!-- Please give a short brief about these changes. -->

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this
change? -->

Fixes