fix(interactive): Fix bug in `ExpandGetVFusionRule`, and optimize when `GetV` has imprecise types  (#3804)
<!--
Thanks for your contribution! please review
https://github.com/alibaba/GraphScope/blob/main/CONTRIBUTING.md before
opening an issue.
-->

## What do these changes do?

<!-- Please give a short brief about these changes. -->

As titled. This pr:
1. Fix a bug in `ExpandGetVFusionRule`. For example, assume we have edge
types of `person-likes-comment`, `person-likes-post`,
`person-knows-person` in schema. Then in queries, if we want to expand
`person-likes-comment`, we would generate a
`Expand(likes)+GetV(comment)` (before this fix, we generate a
`Expand(likes)` only, which is a bug); And if we want to expand
`person-knows-person`, we simply generate a `Expand(knows)`.
2. Optimize cases when the types in GetV is imprecise, to avoid
unnecessary filtering in Runtime.

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this
change? -->

Fixes #3732 #3802

---------

Co-authored-by: Xiaoli Zhou <yihe.zxl@alibaba-inc.com>