feat(interactive): support multiple properties on edge (#3962)
- To support multiple properties on edge, we introduce new type `Record`
and `RecordView` here. `Record` is associated with a property array
through a pointer. `RecordView` consists of an offset and a pointer to a
`Table`.

- Then we specialized a new csr type `mutablecsr<RecordView>` as
multi-properties edge storage. `mutablecsr<RecordView>` contains a
`Table` through which we organize multi-properties. When `get_data()` is
called, a `RecordView` Object will be returned, we can get properties
with the RecordView, and when it is necessary to insert edges into the
graph, we need to encapsulate the multi-attributes into a `Record`.