feat(interactive): Support builtin app in Interactive (#4242)
This PR introduces support for built-in procedures in Interactive. Users
can access these procedures as soon as the graph is created, with more
built-in apps to be added in the future.

For instance, a user can count the vertices of a specific label using:

```cypher
call count_vertices("person");
```

Please note that built-in procedures cannot be deleted or updated, and
users cannot create procedures with the same name as a built-in
procedure.