refactor(interactive): Resolve the dependency problem between compiler and interactive SDK (#4284)
The Interactive SDK and compiler both rely on gaia-ir-proto, including
the protoc-generated classes. Previously, we used the maven-shade-plugin
to shade the protoc files from the Interactive SDK, but this led to
other build issues, as noted in PR #4255.

In this PR, we address the problem by adding an additional jar
classifier, `no-gaia-ir`, which can be easily utilized by the compiler.